Update template Move kover to common update action deprecated
