+++ .github/workflows/ci.yml
@@ -49,7 +49,7 @@ jobs:
         run: ./gradlew :${{ matrix.module }}:build --parallel --build-cache --gradle-user-home ~/.gradle
 
       - name: Archive artifacts
-        uses: actions/upload-artifact@v2
+        uses: actions/upload-artifact@v4
         with:
           name: build-artifacts-${{ matrix.module }}
           path: ${{ matrix.module }}/build/libs/
@@ -83,7 +83,7 @@ jobs:
           cache: gradle
 
       - name: Download build artifacts
-        uses: actions/download-artifact@v2
+        uses: actions/download-artifact@v4
         with:
           name: build-artifacts-${{ matrix.module }}
           path: ${{ matrix.module }}/build/libs/
