+++ build-logic/src/main/kotlin/plugins/CommonMppLibPlugin.kt
@@ -2,6 +2,7 @@ package plugins
 
 import com.android.build.gradle.LibraryExtension
 import extensions.libs
+import extensions.registerPrintLineCoverageTask
 import org.gradle.api.JavaVersion
 import org.gradle.api.Plugin
 import org.gradle.api.Project
@@ -21,7 +22,10 @@ class CommonMppLibPlugin : Plugin<Project> {
                 apply(libs.findPlugin("kotlinPluginSerialization").get().get().pluginId)
                 apply(libs.findPlugin("ktlint").get().get().pluginId)
                 apply(libs.findPlugin("dokka").get().get().pluginId)
+                apply(libs.findPlugin("kover").get().get().pluginId)
             }
+
+            registerPrintLineCoverageTask()
 
             extensions.configure<LibraryExtension> {
                 compileSdk = libs.findVersion("androidCompileSdk").get().displayName.toInt()
